LayoutTests/inspector/workspace-mapping-expected.txt

Tests workspace mappings

    url http://www.example.com/index.html is not mapped.
    url http://localhost/index.html is mapped.
    url http://localhost/foo/index.html is mapped.
    url https://localhost/index.html is not mapped.

    url http://www.example.com/index.html is mapped to http://www.example.com/index.html
    url http://localhost/index.html is mapped to 0@www/localhost/index.html
    url http://localhost/foo/index.html is mapped to null
    url https://localhost/index.html is mapped to https://localhost/index.html

    path /home/example.com/foo/index.html is mapped to http://www.example.com/foo/index.html
    path /home/example.com/index.html is mapped to http://www.example.com/index.html
    path /var/www/localhost/index.html is mapped to http://localhost/index.html
    path /var/www/localhost/foo/index.html is mapped to http://localhost/foo/index.html
    path /home/foo/index.html is mapped to null
